pyproject.toml
[tool.poetry]
name = "ai-image-gen"
version = "0.1.0"
description = "Root dependency tree for image generation experiments"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
diffusers = "^0.34.0"
transformers = "^4.55.0"
safetensors = "^0.6.1"
python-dotenv = "^1.1.1"
accelerate = "^0.34.2"
datasets = "^2.21.0"
peft = ">=0.15.0,<1.0"
pillow = "^10.4.0"
tqdm = "^4.66.4"
numpy = "^2.0.0"
opencv-python-headless = "^4.10.0.84"
requests = "^2.32.3"
huggingface_hub = ">=0.34.0,<1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
